+++ sdk/iothub/azure-resourcemanager-iothub/src/main/java/com/azure/resourcemanager/iothub/implementation/EndpointHealthDataImpl.java
@@ -4,7 +4,6 @@
 
 package com.azure.resourcemanager.iothub.implementation;
 
-import com.azure.resourcemanager.iothub.IotHubManager;
 import com.azure.resourcemanager.iothub.fluent.models.EndpointHealthDataInner;
 import com.azure.resourcemanager.iothub.models.EndpointHealthData;
 import com.azure.resourcemanager.iothub.models.EndpointHealthStatus;
@@ -12,9 +11,10 @@ import com.azure.resourcemanager.iothub.models.EndpointHealthStatus;
 public final class EndpointHealthDataImpl implements EndpointHealthData {
     private EndpointHealthDataInner innerObject;
 
-    private final IotHubManager serviceManager;
+    private final com.azure.resourcemanager.iothub.IotHubManager serviceManager;
 
-    EndpointHealthDataImpl(EndpointHealthDataInner innerObject, IotHubManager serviceManager) {
+    EndpointHealthDataImpl(
+        EndpointHealthDataInner innerObject, com.azure.resourcemanager.iothub.IotHubManager serviceManager) {
         this.innerObject = innerObject;
         this.serviceManager = serviceManager;
     }
@@ -31,7 +31,7 @@ public final class EndpointHealthDataImpl implements EndpointHealthData {
         return this.innerObject;
     }
 
-    private IotHubManager manager() {
+    private com.azure.resourcemanager.iothub.IotHubManager manager() {
         return this.serviceManager;
     }
 }