+++ sdk/iothub/azure-resourcemanager-iothub/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0-beta.2 (2021-05-07)
 
+- Azure Resource Manager IotHub client library for Java. This package contains Microsoft Azure SDK for IotHub Management SDK. Use this API to manage the IoT hubs in your Azure subscription. Package tag package-2020-03. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
+
+### Breaking Change
+
+#### `models.CertificateDescription$Update` was modified
+
+* `withIfMatch(java.lang.String)` was removed
+
+### New Feature
+
+#### `models.CertificateDescription$Update` was modified
+
+* `ifMatch(java.lang.String)` was added
 
 ## 1.0.0-beta.1 (2021-03-02)